Move defaultIncomeTypes to constants
