+++ libs/application/templates/social-insurance-administration/core/src/lib/socialInsuranceAdministrationUtils.ts
@@ -204,41 +204,3 @@ export const getTypesOptions = (
       }
     })
 }
-
-export const defaultIncomeTypes = [
-  {
-    income: 'yearly',
-    currency: 'IKR',
-    incomeType: 'Lífeyrissjóður',
-    incomePerYear: '0',
-    incomeCategory: 'Lífeyrissjóðstekjur',
-  },
-  {
-    income: 'yearly',
-    currency: 'IKR',
-    incomeType: 'Laun',
-    incomePerYear: '0',
-    incomeCategory: 'Atvinnutekjur',
-  },
-  {
-    income: 'yearly',
-    currency: 'IKR',
-    incomeType: 'Vextir af innistæðum',
-    incomePerYear: '0',
-    incomeCategory: 'Fjármagnstekjur',
-  },
-  {
-    income: 'yearly',
-    currency: 'EUR',
-    incomeType: 'Erlendur lífeyrir',
-    incomePerYear: '0',
-    incomeCategory: 'Lífeyrissjóðstekjur',
-  },
-  {
-    income: 'yearly',
-    currency: 'IKR',
-    incomeType: 'Vextir af verðbréfum',
-    incomePerYear: '0',
-    incomeCategory: 'Fjármagnstekjur',
-  },
-]
